started refactoring input
