+++ src/input/unix_input.rs
@@ -1,21 +1,20 @@
 //! This is a UNIX specific implementation for input related action.
 
-use std::char;
 use std::sync::{
+    char,
+    io::{self, Read},
+    str, thread,
     atomic::{AtomicBool, Ordering},
     mpsc::{self, Receiver, Sender},
     Arc,
-};
-use std::{
-    io::{self, Read},
-    str, thread,
 };
 
 use crossterm_utils::{csi, write_cout, ErrorKind, Result};
 
 use crate::sys::unix::{get_tty, read_char_raw};
 
-use super::{ITerminalInput, InputEvent, KeyEvent, MouseButton, MouseEvent};
+use crate::input::ITerminalInput;
+use crate::{InputEvent, KeyEvent, MouseButton, MouseEvent};
 
 pub struct UnixInput;
